Move ChatHIstory to schema to kill circular dependency
