--- src/steamship/agents/memory/__init__.py
@@ -1,3 +1,0 @@
-from .chathistory import ChatHistory
-
-__all__ = ["ChatHistory"]
