+++ src/steamship/agents/schema/__init__.py
@@ -1,5 +1,6 @@
 from .action import Action, FinishAction
 from .agent import Agent, LLMAgent
+from .chathistory import ChatHistory
 from .context import AgentContext, EmitFunc, Metadata
 from .llm import LLM
 from .output_parser import OutputParser
@@ -16,4 +17,5 @@ __all__ = [
     "LLMAgent",
     "OutputParser",
     "Tool",
+    "ChatHistory",
 ]
